feat(arena) : created the arean feature

+++ src/http/modules/space/space.service.ts
@@ -9,36 +9,65 @@ export const SpaceService = {
     mapId: string,
     creatorId: string,
   ) {
-    const space = await prisma.space.create({
-      data: {
-        name: name,
-        height: parseInt(dimensions.split("x")[0]!),
-        width: parseInt(dimensions.split("x")[1]!),
-        creatorId: creatorId,
-        mapId: mapId,
+    const map = await prisma.map.findUnique({
+      where: {
+        id: mapId,
       },
     });
+    if (!map) {
+      throw "Map Do not exist";
+    } else {
+      const space = await prisma.space.create({
+        data: {
+          name: name,
+          height: parseInt(dimensions.split("x")[0]!),
+          width: parseInt(dimensions.split("x")[1]!),
+          creatorId: creatorId,
+          mapId: mapId,
+        },
+      });
 
-    return space.id;
+      return space.id;
+    }
   },
 
   //delete a space
 
   async deleteSpace(spaceId: string) {
-    await prisma.space.delete({
+    const space = await prisma.space.findUnique({
       where: {
         id: spaceId,
       },
     });
+
+    if (!space) {
+      throw "Space do not exist";
+    } else {
+      await prisma.space.delete({
+        where: {
+          id: spaceId,
+        },
+      });
+    }
   },
 
   // get my existing space
 
   async getMySpaces(userId: string) {
-    return await prisma.space.findMany({
+    const user = await prisma.user.findUnique({
       where: {
-        creatorId: userId,
+        id: userId,
       },
     });
+
+    if (!user) {
+      throw Error("User do not exist");
+    } else {
+      return await prisma.space.findMany({
+        where: {
+          creatorId: userId,
+        },
+      });
+    }
   },
 };
